requirements.txt

fastapi
uvicorn
sqlalchemy
python-jose
passlib
jinja2
